refactor(vendor): user smux in depends to replace pkg in vendor. #1000193713
