+++ depends/xtaci/smux/alloc.go
