+++ metanode/server.go
@@ -20,7 +20,7 @@ import (
 	"net"
 	"strings"
 
-	"github.com/xtaci/smux"
+	"github.com/cubefs/cubefs/depends/xtaci/smux"
 
 	"github.com/cubefs/cubefs/proto"
 	"github.com/cubefs/cubefs/util"
